+++ app/code/local/Mulberry/Warranty/Model/Observer/Api.php
@@ -10,23 +10,6 @@
 class Mulberry_Warranty_Model_Observer_Api
 {
     /**
-     * Send order information to Mulberry
-     *
-     * @param Varien_Event_Observer $observer
-     */
-    public function sendOrder(Varien_Event_Observer $observer)
-    {
-        /**
-         * @var Mage_Sales_Model_Order $order
-         */
-        $order = $observer->getEvent()->getOrder();
-
-        if (Mage::helper('mulberry_warranty')->isActive()) {
-            Mage::getModel('mulberry_warranty/api_rest_send_order')->sendOrder($order);
-        }
-    }
-
-    /**
      * Perform order cancellation for Mulberry warranty products
      *
      * @param Varien_Event_Observer $observer
@@ -40,23 +23,6 @@ class Mulberry_Warranty_Model_Observer_Api
 
         if (Mage::helper('mulberry_warranty')->isActive()) {
             Mage::getModel('mulberry_warranty/api_rest_cancel_order')->cancelOrder($order);
-        }
-    }
-
-    /**
-     * Send order information to Mulberry
-     *
-     * @param Varien_Event_Observer $observer
-     */
-    public function sendCart(Varien_Event_Observer $observer)
-    {
-        /**
-         * @var Mage_Sales_Model_Order $order
-         */
-        $order = $observer->getEvent()->getOrder();
-
-        if (Mage::helper('mulberry_warranty')->isActive()) {
-            Mage::getModel('mulberry_warranty/api_rest_send_cart')->sendCart($order);
         }
     }